requirements.txt
appdirs==1.4.3
attrs==19.3.0
backcall==0.1.0
black==19.10b0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
decorator==4.4.1
distlib==0.3.0
entrypoints==0.3
filelock==3.0.12
flake8==3.7.9
idna==2.9
importlib-metadata==1.5.0
ipython==7.12.0
ipython-genutils==0.2.0
jedi==0.16.0
mccabe==0.6.1
packaging==20.1
parso==0.6.1
pathspec==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
pluggy==0.13.1
prompt-toolkit==3.0.3
ptyprocess==0.6.0
py==1.8.1
pycodestyle==2.5.0
pyflakes==2.1.1
Pygments==2.5.2
pyparsing==2.4.6
python-frontmatter==0.5.0
PyYAML==5.3
regex==2020.2.20
requests==2.23.0
rope==0.16.0
selenium==3.141.0
six==1.14.0
toml==0.10.0
tox==3.14.5
traitlets==4.3.3
typed-ast==1.4.1
urllib3==1.25.8
virtualenv==20.0.7
wcwidth==0.1.8
webbot==0.1.4
zipp==3.0.0
